requirements.txt
flask~=2.1.3
twilio~=6.38.1
faker~=4.0.3
python-dotenv==0.13.0
pytest==5.4.1
